+++ src/common/mod.rs
@@ -2,5 +2,6 @@
 
 pub mod bool_or;
 pub mod extensions;
+pub mod formats;
 pub mod helpers;
 pub mod reference;
