+++ src/v2/items.rs
@@ -5,8 +5,8 @@ use std::ops::Add;
 
 use serde::{Deserialize, Serialize};
 
+use crate::common::formats::{CollectionFormat, IntegerFormat, NumberFormat, StringFormat};
 use crate::common::helpers::{Context, ValidateWithContext};
-use crate::v2::formats::{CollectionFormat, IntegerFormat, NumberFormat, StringFormat};
 use crate::v2::spec::Spec;
 
 #[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
