+++ src/v2/parameter.rs
@@ -5,9 +5,9 @@ use std::ops::Add;
 
 use serde::{Deserialize, Serialize};
 
+use crate::common::formats::{CollectionFormat, IntegerFormat, NumberFormat, StringFormat};
 use crate::common::helpers::{validate_required_string, Context, ValidateWithContext};
 use crate::common::reference::RefOr;
-use crate::v2::formats::{CollectionFormat, IntegerFormat, NumberFormat, StringFormat};
 use crate::v2::items::Items;
 use crate::v2::schema::Schema;
 use crate::v2::spec::Spec;
